requirements.txt
torch
numpy
pyyaml
msgpack
msgpack-numpy
lz4
pymongo
